add validateData
