+++ controllers/contacts.js
@@ -1,6 +1,6 @@
 const contacts = require("../models/contacts");
 const { HttpError } = require("../helpers");
-const schemas = require("../schemas/contacts");
+// const schemas = require("../schemas/contacts");
 
 const getAllContacts = async (req, res, next) => {
   try {
@@ -24,28 +24,37 @@ const getContact = async (req, res, next) => {
   }
 };
 
+// const addNewContact = async (req, res, next) => {
+//   try {
+//     const { error } = schemas.addSchema.validate(req.body);
+//     const emptyBody = Object.keys(req.body).length === 0;
+//     const missingField =
+//       schemas.addSchema._ids._byKey.size !== Object.keys(req.body).length;
+
+//     if (emptyBody) {
+//       throw HttpError(400, "missing fields");
+//     }
+
+//     if (error) {
+//       if (missingField) {
+//         throw HttpError(
+//           400,
+//           `missing required ${error.details[0].path[0]} field`
+//         );
+//       } else {
+//         throw HttpError(400, error.message);
+//       }
+//     }
+
+//     const result = await contacts.addContact(req.body);
+//     res.status(201).json(result);
+//   } catch (error) {
+//     next(error);
+//   }
+// };
+
 const addNewContact = async (req, res, next) => {
   try {
-    const { error } = schemas.addSchema.validate(req.body);
-    const emptyBody = Object.keys(req.body).length === 0;
-    const missingField =
-      schemas.addSchema._ids._byKey.size !== Object.keys(req.body).length;
-
-    if (emptyBody) {
-      throw HttpError(400, "missing fields");
-    }
-
-    if (error) {
-      if (missingField) {
-        throw HttpError(
-          400,
-          `missing required ${error.details[0].path[0]} field`
-        );
-      } else {
-        throw HttpError(400, error.message);
-      }
-    }
-
     const result = await contacts.addContact(req.body);
     res.status(201).json(result);
   } catch (error) {
@@ -68,28 +77,41 @@ const delContact = async (req, res, next) => {
   }
 };
 
+// const editContact = async (req, res, next) => {
+//   try {
+//     const { error } = schemas.addSchema.validate(req.body);
+//     const emptyBody = Object.keys(req.body).length === 0;
+//     const missingField =
+//       schemas.addSchema._ids._byKey.size !== Object.keys(req.body).length;
+
+//     if (emptyBody) {
+//       throw HttpError(400, "missing fields");
+//     }
+
+//     if (error) {
+//       if (missingField) {
+//         throw HttpError(
+//           400,
+//           `missing required ${error.details[0].path[0]} field`
+//         );
+//       } else {
+//         throw HttpError(400, error.message);
+//       }
+//     }
+
+//     const { contactId } = req.params;
+//     const result = await contacts.updateContact(contactId, req.body);
+//     if (!result) {
+//       throw HttpError(404, "Not found");
+//     }
+//     res.json(result);
+//   } catch (error) {
+//     next(error);
+//   }
+// };
+
 const editContact = async (req, res, next) => {
   try {
-    const { error } = schemas.addSchema.validate(req.body);
-    const emptyBody = Object.keys(req.body).length === 0;
-    const missingField =
-      schemas.addSchema._ids._byKey.size !== Object.keys(req.body).length;
-
-    if (emptyBody) {
-      throw HttpError(400, "missing fields");
-    }
-
-    if (error) {
-      if (missingField) {
-        throw HttpError(
-          400,
-          `missing required ${error.details[0].path[0]} field`
-        );
-      } else {
-        throw HttpError(400, error.message);
-      }
-    }
-
     const { contactId } = req.params;
     const result = await contacts.updateContact(contactId, req.body);
     if (!result) {
