+++ middlewares/index.js
@@ -1,0 +1,5 @@
+const validateData = require("./validateData");
+
+module.exports = {
+  validateData,
+};
