+++ models/contacts/contacts.json
@@ -1,7 +1,7 @@
 [
   {
     "id": "qdggE76Jtbfd9eWJHrssH",
-    "name": "nameNew1",
+    "name": "nameNew2",
     "email": "emailNew",
     "phone": "phoneNew"
   },
@@ -56,6 +56,12 @@
   {
     "id": "Tt31dHV4ModyqKHlc_N1l",
     "name": "nameNew1",
+    "email": "emailNew",
+    "phone": "phoneNew"
+  },
+  {
+    "id": "Tt3LXJepMqHF5xNEWNn_g",
+    "name": "nameNew2",
     "email": "emailNew",
     "phone": "phoneNew"
   }
